refactor: move static routers to a dedicated file
