+++ test/test_api.py
@@ -10,30 +10,6 @@ def test_get_root():
     assert response.json() == {"Hello": "World"}
 
 
-def test_get_version_txt():
-    response = client.get("/version.txt")
-    assert response.status_code == 200
-    assert response.headers.get("content-type") == "text/plain; charset=utf-8"
-    assert response.text == "# It will be replaced by docker build\n"
-
-
-def test_get_favicon():
-    response = client.get("/favicon.ico")
-    assert response.status_code == 200
-    assert response.headers.get("content-type") in {
-        # Mac hosts
-        "image/x-icon",
-        # Linux hosts
-        "image/vnd.microsoft.icon",
-    }
-
-
-def test_get_robots_txt():
-    response = client.get("/robots.txt")
-    assert response.status_code == 200
-    assert response.headers.get("content-type") == "text/plain; charset=utf-8"
-
-
 def test_get_openapi_json():
     response = client.get("/openapi.json")
     assert response.status_code == 200
